requirements.txt
openai
pymongo
python_dotenv